backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
firebase-admin==6.5.0
psycopg2-binary 
